compile styles in main page components & rename component Help to MainPage & transfer component Splash from view folder to features
