+++ src/components/common/ProgramName/ProgramName.js
@@ -3,7 +3,7 @@ import styles from '../ProgramName/ProgramName.module.scss';
 
 const ProgramName = () => (
   <div className={styles.root}>
-    <h1 className={styles.programName}>program operacyjny pomoc żywnościowa 2014-2020</h1>
+    <h1 className={styles.programName}>pomoc żywnościowa 2014-2020</h1>
     <h2 className={styles.podprogramName}>podprogram 2021</h2>
   </div>
 );
