+++ src/components/features/Splash/Splash.module.scss
