+++ src/components/layout/Header/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './Header.module.scss';
 import  Navigation from '../Navigation/Navigation';
-import  Splash  from '../../views/Splash/Splash';
+import Splash from '../../features/Splash/Splash';
 
 const Header = () => (
   <header className={styles.header}>
