--- src/components/views/Help/Help.module.scss
@@ -1,14 +1,0 @@
-@import '../../../styles/global.scss';
-
-.root {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: $text-color;
-  padding: 20px;
-  margin: 15px 200px 15px 200px;
-
-  input {
-    width: 80vw;
-  }
-}
